+++ CHANGES.rst
@@ -6,7 +6,10 @@ Changelog
 ------------------
 
 - Reset File object actions on uninstall [gyst]
+
 - Propagate UI language setting to Collabora [gyst]
+
+- Update Github CI infrastructure [gyst]
 
 
 1.1.0 (2025-07-25)